Update widget to use a callback to set deckJson
