+++ modules/jupyter-widget/src/index.js
@@ -17,4 +17,5 @@ if (dataBaseUrl) {
 }
 
 export {MODULE_VERSION, MODULE_NAME} from './version';
-export {DeckGLModel, DeckGLView, initDeck} from './widget';
+export {DeckGLModel, DeckGLView} from './widget';
+export {initDeck} from './utils';
